// serviceLayer/src/fbs/generated/model/iLLBibliographicRecord.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * FBS Adapter
 * OpenAPI spec version: 1.0
 */

export interface ILLBibliographicRecord {
  /** The author of the material */
  author?: string;
  /** Bibliographic category from danMARC2 008 *t */
  bibliographicCategory?: string;
  /** Edition-information from the bibliographic record */
  edition?: string;
  /** ISBN-information from the bibliographic record */
  isbn?: string;
  /** ISSN-information from the bibliographic record */
  issn?: string;
  /** Language of the requested material. */
  language?: string;
  /** Type of the requested material - from danMARC2 009 *a+*g (general and specific) */
  mediumType: string;
  /** Issue number of a periodical */
  periodicalNumber?: string;
  /** Volume name of a periodical */
  periodicalVolume?: string;
  placeOfPublication?: string;
  /** Publication date of the requested material. */
  publicationDate?: string;
  /** Publication date of an item component, or article. */
  publicationDateOfComponent?: string;
  /** Publisher of the requested material. */
  publisher?: string;
  /** The FAUST number */
  recordId: string;
  /** The title of the material */
  title?: string;
}