// serviceLayer/src/fbs/generated/model/loanDetailsV2.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * FBS Adapter
 * OpenAPI spec version: 1.0
 */
import type { ILLBibliographicRecord } from "./iLLBibliographicRecord";
import type { MaterialGroup } from "./materialGroup";
import type { Periodical } from "./periodical";

export interface LoanDetailsV2 {
  /** The date when the material must be returned */
  dueDate: string;
  ilBibliographicRecord?: ILLBibliographicRecord;
  /** The date when the material was picked up */
  loanDate: string;
  /** Identifies the loan for use when renewing the loan */
  loanId: number;
  /** The loan type, either <b>loan</b> or <b>interLibraryLoan</b> */
  loanType: string;
  materialGroup: MaterialGroup;
  /** Identifies the exact material that has been loaned */
  materialItemNumber: string;
  periodical?: Periodical;
  /** The FAUST number of the bibliographic record */
  recordId: string;
}