// serviceLayer/src/fbs/generated/model/renewedLoanV2.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * FBS Adapter
 * OpenAPI spec version: 1.0
 */
import type { LoanDetailsV2 } from "./loanDetailsV2";

export interface RenewedLoanV2 {
  loanDetails: LoanDetailsV2;
  /** indicates if renewal was succesful or denied - including the reason for denial. */
  renewalStatus: string[];
}